# Improving performance of assembling routines in `DiffusionPWLCSolver` and `DiffusionMIPSolver`