This PR improves the assembling routines in `DiffusionPWLCSolver` and `DiffusionMIPSolver`:

- the dof map is computed just once
- the local contributions for the matrix and right-hand side vectors are accumulated into a dense matrix/vector, then added to the global matrix, vector

Impact observed on the C5G7 benchmark:

### `AssembleAand_b` routine:

| Case | `main` | `optimized` | Speedup |
| --- | ---: | ---: | ---:|
| `DiffusionMIPSolver` | 11.189 | 5.619 | 1.9915 |
| `DiffusionPWLCSolver` | 1.422 | 0.919 | 1.5467 |

### `Assemble_b` routine:

| Case | `main` | `optimized` | Speedup |
| --- | ---: | ---: | ---:|
| `DiffusionMIPSolver` | 0.170 | 0.153 | 1.1114 |
| `DiffusionPWLCSolver` | 0.238 | 0.223 | 1.0654 |

All times are in seconds.