@ragusa 2024-11-16
> > `assembleAand_b` and `assemble_b` do not show symmetry in their names. should we change the first one to `assemble_A_and_b` ?
> 
> I think we talked about this once before, but no decision was made. Strictly speaking, the names should be `AssembleAAndB` and `AssembleB`, but capital `B` would suggest `b` might be a matrix. That's probably why there is `_` in the name. Alternatively `AssembleMatrixAndRhs` and `AssembleRhs` - we use `PascalCase` for function names...

ok. I could be fine with this.